fix(ms2/modeler): don't open xml on ctl+x when a modal is open
