+++ src/management-system-v2/app/(dashboard)/[environmentId]/processes/[processId]/modeler-toolbar.tsx
@@ -1,4 +1,4 @@
-import React, { ComponentProps, use, useEffect, useState } from 'react';
+import { ComponentProps, use, useEffect, useState } from 'react';
 import { is as bpmnIs } from 'bpmn-js/lib/util/ModelUtil';
 import { App, Tooltip, Button, Space, Select, SelectProps, Divider } from 'antd';
 import { Toolbar, ToolbarGroup } from '@/components/toolbar';
@@ -34,16 +34,17 @@ import { EnvVarsContext } from '@/components/env-vars-context';
 import { Process } from '@/lib/data/process-schema';
 import FlowConditionModal, { isConditionalFlow } from './flow-condition-modal';
 import { TimerEventButton, isTimerEvent } from './planned-duration-input';
+import XmlEditor from './xml-editor';
 
 const LATEST_VERSION = { id: '-1', name: 'Latest Version', description: '' };
 
 type ModelerToolbarProps = {
   process: Process;
-  onOpenXmlEditor: () => void;
   canUndo: boolean;
   canRedo: boolean;
+  versionName?: string;
 };
-const ModelerToolbar = ({ process, onOpenXmlEditor, canUndo, canRedo }: ModelerToolbarProps) => {
+const ModelerToolbar = ({ process, canRedo, canUndo, versionName }: ModelerToolbarProps) => {
   const processId = process.id;
 
   const router = useRouter();
@@ -75,13 +76,47 @@ const ModelerToolbar = ({ process, onOpenXmlEditor, canUndo, canRedo }: ModelerT
   // Force rerender when the BPMN changes.
   useModelerStateStore((state) => state.changeCounter);
 
+  const [xmlEditorBpmn, setXmlEditorBpmn] = useState<string | undefined>(undefined);
+  const handleOpenXmlEditor = async () => {
+    // Undefined can maybe happen when click happens during router transition?
+    if (modeler) {
+      const xml = await modeler.getXML();
+      setXmlEditorBpmn(xml);
+    }
+  };
+  const handleXmlEditorSave = async (bpmn: string) => {
+    if (modeler) {
+      await modeler.loadBPMN(bpmn);
+      // If the bpmn contains unexpected content (text content for an element
+      // where the model does not define text) the modeler will remove it
+      // automatically => make sure the stored bpmn is the same as the one in
+      // the modeler.
+      const cleanedBpmn = await modeler.getXML();
+      await updateProcess(process.id, environment.spaceId, cleanedBpmn);
+    }
+  };
+
   const modalOpen =
-    showUserTaskEditor || showPropertiesPanel || showScriptTaskEditor || shareModalOpen;
+    showUserTaskEditor ||
+    showPropertiesPanel ||
+    showScriptTaskEditor ||
+    shareModalOpen ||
+    !!xmlEditorBpmn;
+
   useEffect(() => {
     if (modalOpen) {
       modeler?.deactivateKeyboard();
     } else modeler?.activateKeyboard();
   }, [modeler, modalOpen]);
+
+  useAddControlCallback(
+    'modeler',
+    'cut',
+    () => {
+      if (!modalOpen) handleOpenXmlEditor();
+    },
+    { blocking: modalOpen },
+  );
 
   const selectedVersionId = query.get('version');
 
@@ -301,7 +336,7 @@ const ModelerToolbar = ({ process, onOpenXmlEditor, canUndo, canRedo }: ModelerT
                 <Tooltip title="Show XML">
                   <Button
                     icon={<Icon aria-label="xml-sign" component={SvgXML} />}
-                    onClick={onOpenXmlEditor}
+                    onClick={handleOpenXmlEditor}
                   />
                 </Tooltip>
               )}
@@ -349,12 +384,23 @@ const ModelerToolbar = ({ process, onOpenXmlEditor, canUndo, canRedo }: ModelerT
           </Space>
         </Space>
       </Toolbar>
+
       <ShareModal
         processes={[process]}
         open={shareModalOpen}
         setOpen={setShareModalOpen}
         defaultOpenTab={shareModalDefaultOpenTab}
       />
+
+      <XmlEditor
+        bpmn={xmlEditorBpmn}
+        canSave={!selectedVersionId}
+        onClose={() => setXmlEditorBpmn(undefined)}
+        onSaveXml={handleXmlEditorSave}
+        process={process}
+        versionName={versionName}
+      />
+
       {env.PROCEED_PUBLIC_ENABLE_EXECUTION && (
         <>
           <UserTaskBuilder
